requirements.txt
numpy
pandas
matplotlib
scikit-learn
python_dotenv
mysql-connector_python
pymysql
seaborn
xgboost
mlflow
catboost


-e .